Allow all user csv downloads to have a 'registered on' column. (#527) (#533)
* Remove db:migration files that are causing migration errors.

These migrations fail to run on our existing database. This is because 
1) our users table already has an index 
2) there exist duplicate names in the db, so unique index cannot be generated.

To fix this we will first remove these two migration files. No need to rollback because the migration failed in the first place (correct me if I am wrong) and nothing happened to the db. Thanks!

* Allow all user csv downloads to have a 'registered on' column.

* Run the correct, updated test to test students.csv output with 'Registered on' column.

The test previously did not validate the new, added column of the csv and was failing.