Fix forum notifications, part 2 (#536)
* Turned off moderation and forced index on users.

* Added email headers for notifications.

* Removed Thredded from email handler. Our app should deal with it.

* Remove Thredded notifications for good.

The logs still indicate Thredded::Notifier is being called. Uncommented a line in config/initializers/thredded.rb which would force it to not be called.

* Add an email from parameter to Thredded's notifier system.

Strangley the system is still looking for a from address, even though notifications have been disabled. We will see if the same error shows up again on post.

* Disable automatic following of forum posts and topics.

* Reactivate forum notifier system.

This will prepare us to present our case to the Thredded issue tracker.